inicia modelagem de arquitetura msc
